fix fro mrg32k

+++ cpp/oneapi/dal/backend/primitives/rng/rng_dpc.cpp
@@ -148,15 +148,15 @@ INSTANTIATE_FLOAT_CPU(std::int32_t);
 #define INSTANTIATE_UNIFORM_WITHOUT_REPLACEMENT_GPU_FLOAT(Size)                     \
     INSTANTIATE_UNIFORM_WITHOUT_REPLACEMENT_GPU(float, Size, engine_list::mt2203)   \
     INSTANTIATE_UNIFORM_WITHOUT_REPLACEMENT_GPU(float, Size, engine_list::mcg59)    \
-    INSTANTIATE_UNIFORM_WITHOUT_REPLACEMENT_GPU(float, Size, engine_list::mrg32k3a)    \
+    INSTANTIATE_UNIFORM_WITHOUT_REPLACEMENT_GPU(float, Size, engine_list::mrg32k3a) \
     INSTANTIATE_UNIFORM_WITHOUT_REPLACEMENT_GPU(float, Size, engine_list::mt19937)  \
     INSTANTIATE_UNIFORM_WITHOUT_REPLACEMENT_GPU(double, Size, engine_list::mt2203)  \
     INSTANTIATE_UNIFORM_WITHOUT_REPLACEMENT_GPU(double, Size, engine_list::mcg59)   \
-    INSTANTIATE_UNIFORM_WITHOUT_REPLACEMENT_GPU(float, Size, engine_list::mrg32k3a)    \
+    INSTANTIATE_UNIFORM_WITHOUT_REPLACEMENT_GPU(double, Size, engine_list::mrg32k3a) \
     INSTANTIATE_UNIFORM_WITHOUT_REPLACEMENT_GPU(double, Size, engine_list::mt19937) \
     INSTANTIATE_UNIFORM_WITHOUT_REPLACEMENT_GPU(int, Size, engine_list::mt2203)     \
     INSTANTIATE_UNIFORM_WITHOUT_REPLACEMENT_GPU(int, Size, engine_list::mcg59)      \
-    INSTANTIATE_UNIFORM_WITHOUT_REPLACEMENT_GPU(float, Size, engine_list::mrg32k3a)    \
+    INSTANTIATE_UNIFORM_WITHOUT_REPLACEMENT_GPU(int, Size, engine_list::mrg32k3a) \
     INSTANTIATE_UNIFORM_WITHOUT_REPLACEMENT_GPU(int, Size, engine_list::mt19937)
 
 INSTANTIATE_UNIFORM_WITHOUT_REPLACEMENT_GPU_FLOAT(std::int64_t);
@@ -170,10 +170,10 @@ INSTANTIATE_UNIFORM_WITHOUT_REPLACEMENT_GPU_FLOAT(std::int32_t);
         oneapi_engine<EngineType>& engine_,                       \
         const event_vector& deps);
 
-#define INSTANTIATE_SHUFFLE_FLOAT(Size)                 \
-    INSTANTIATE_SHUFFLE(int, Size, engine_list::mt2203) \
-    INSTANTIATE_SHUFFLE(int, Size, engine_list::mcg59)  \
-    INSTANTIATE_SHUFFLE(int, Size, engine_list::mrg32k3a)  \
+#define INSTANTIATE_SHUFFLE_FLOAT(Size)                   \
+    INSTANTIATE_SHUFFLE(int, Size, engine_list::mt2203)   \
+    INSTANTIATE_SHUFFLE(int, Size, engine_list::mcg59)    \
+    INSTANTIATE_SHUFFLE(int, Size, engine_list::mrg32k3a) \
     INSTANTIATE_SHUFFLE(int, Size, engine_list::mt19937)
 
 INSTANTIATE_SHUFFLE_FLOAT(std::int64_t);
